refactor: enable use of external client library or autogen client
Add an adapter WESClient class to map workflow-service client
functions to controllers in the WES spec. The WES class now
serves as the primary Python interface for WES functionality
in the orchestrator, and can use either workflow-service or
the autogenerated client (bravado) to make API calls to an
endpoint. This combines my earlier changes in #30 with some
helpful additions and improvements from @DailyDreaming in
later commits.

See #27