fix: modify bravado TRS client to use correct api url
Use the bravado loader class to construct a spec dict from the
swagger yaml, then load the client from the spec with the proper
origin url specified. Loading 'from_url' with a local file
doesn't provide an option to specify the origin url and instead
uses the relative local path (even with the http client provided).